move asset folders to top-level in output
Move css, js, and fonts folders to top-level folder in the final output,
rather than being grouped in an "assets" folder.  This is closer to what
Hugo expects, and therefore allows the asset pipeline to be used a
little more naturally (without needing to override a handful of values),
and makes it a little cleaner to support custom css.